# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Tax Sathi API for India income-tax filing preparation
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DocumentDocumentType = typeof DocumentDocumentType[keyof typeof DocumentDocumentType];


export const DocumentDocumentType = {
  form16: 'form16',
  form26as: 'form26as',
  ais: 'ais',
  rent_receipt: 'rent_receipt',
  investment_proof: 'investment_proof',
} as const;

export type DocumentStatus = typeof DocumentStatus[keyof typeof DocumentStatus];


export const DocumentStatus = {
  processing: 'processing',
  review: 'review',
  verified: 'verified',
} as const;

export interface ExtractedFields {
  employer: string;
  /** @minimum 0 */
  grossSalary: number;
  /** @minimum 0 */
  tdsDeducted: number;
  pan: string;
  assessmentYear: string;
  confidenceNote?: string;
}

export interface Document {
  id: string;
  fileName: string;
  documentType: DocumentDocumentType;
  status: DocumentStatus;
  uploadedAt: string;
  confidence: number;
  extracted: ExtractedFields;
}

export type DocumentInputDocumentType = typeof DocumentInputDocumentType[keyof typeof DocumentInputDocumentType];


export const DocumentInputDocumentType = {
  form16: 'form16',
  form26as: 'form26as',
  ais: 'ais',
  rent_receipt: 'rent_receipt',
  investment_proof: 'investment_proof',
} as const;

export interface DocumentInput {
  /** @minLength 1 */
  fileName: string;
  documentType: DocumentInputDocumentType;
  /** @minimum 1 */
  pageCount?: number;
}

export type DocumentUpdateStatus = typeof DocumentUpdateStatus[keyof typeof DocumentUpdateStatus];


export const DocumentUpdateStatus = {
  processing: 'processing',
  review: 'review',
  verified: 'verified',
} as const;

export interface DocumentUpdate {
  status?: DocumentUpdateStatus;
  extracted?: ExtractedFields;
}

export interface UserSummary {
  name: string;
  panMasked: string;
  filingType: string;
}

export interface RegimeResult {
  taxableIncome: number;
  baseTax: number;
  cess: number;
  totalTax: number;
  effectiveRate: number;
  rebateApplied?: boolean;
  deductions?: number;
}

export type TaxSummaryRecommendedRegime = typeof TaxSummaryRecommendedRegime[keyof typeof TaxSummaryRecommendedRegime];


export const TaxSummaryRecommendedRegime = {
  new: 'new',
  old: 'old',
} as const;

export interface TaxSummary {
  grossIncome: number;
  newRegime: RegimeResult;
  oldRegime: RegimeResult;
  recommendedRegime: TaxSummaryRecommendedRegime;
  savings: number;
  calculatedAt: string;
}

export type ChecklistItemState = typeof ChecklistItemState[keyof typeof ChecklistItemState];


export const ChecklistItemState = {
  complete: 'complete',
  current: 'current',
  upcoming: 'upcoming',
} as const;

export interface ChecklistItem {
  id: string;
  label: string;
  state: ChecklistItemState;
}

export interface Dashboard {
  user: UserSummary;
  assessmentYear: string;
  documents: Document[];
  tax: TaxSummary;
  checklist: ChecklistItem[];
}

export interface TaxCalculationInput {
  /** @minimum 0 */
  grossSalary: number;
  /** @minimum 0 */
  otherIncome?: number;
  /** @minimum 0 */
  deductions80c?: number;
  /** @minimum 0 */
  deductions80d?: number;
  /** @minimum 0 */
  deductions80g?: number;
  /** @minimum 0 */
  homeLoanInterest?: number;
  /** @minimum 0 */
  hraReceived?: number;
  /** @minimum 0 */
  rentPaid?: number;
  /** @minimum 0 */
  basicSalary?: number;
  isMetro?: boolean;
  /** @minimum 0 */
  age?: number;
}

export type ItrDraftInputRegime = typeof ItrDraftInputRegime[keyof typeof ItrDraftInputRegime];


export const ItrDraftInputRegime = {
  new: 'new',
  old: 'old',
} as const;

export interface ItrDraftInput {
  /** @pattern ^[A-Z]{5}[0-9]{4}[A-Z]$ */
  pan: string;
  /** @minLength 1 */
  name: string;
  /** @minimum 0 */
  grossSalary: number;
  /** @minimum 0 */
  tdsDeducted: number;
  regime: ItrDraftInputRegime;
}

export type ItrDraftForm = typeof ItrDraftForm[keyof typeof ItrDraftForm];


export const ItrDraftForm = {
  'ITR-1': 'ITR-1',
} as const;

export type ItrDraftStatus = typeof ItrDraftStatus[keyof typeof ItrDraftStatus];


export const ItrDraftStatus = {
  draft: 'draft',
  needs_review: 'needs_review',
  ready: 'ready',
} as const;

export type ItrDraftPayload = { [key: string]: unknown };

export interface ItrDraft {
  id: string;
  form: ItrDraftForm;
  assessmentYear: string;
  status: ItrDraftStatus;
  generatedAt: string;
  warnings: string[];
  payload: ItrDraftPayload;
}

export interface Consent {
  ocrProcessing: boolean;
  dataStorage: boolean;
  taxCalculation: boolean;
  updatedAt: string;
}

export interface ConsentInput {
  ocrProcessing: boolean;
  dataStorage: boolean;
  taxCalculation: boolean;
}