# lib/api-zod/src/generated/types/checklistItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Tax Sathi API for India income-tax filing preparation
 * OpenAPI spec version: 0.1.0
 */
import type { ChecklistItemState } from './checklistItemState';

export interface ChecklistItem {
  id: string;
  label: string;
  state: ChecklistItemState;
}