# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Tax Sathi API for India income-tax filing preparation
 * OpenAPI spec version: 0.1.0
 */
import type { ChecklistItem } from './checklistItem';
import type { Document } from './document';
import type { TaxSummary } from './taxSummary';
import type { UserSummary } from './userSummary';

export interface Dashboard {
  user: UserSummary;
  assessmentYear: string;
  documents: Document[];
  tax: TaxSummary;
  checklist: ChecklistItem[];
}