# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Tax Sathi API for India income-tax filing preparation
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentDocumentType } from './documentDocumentType';
import type { DocumentStatus } from './documentStatus';
import type { ExtractedFields } from './extractedFields';

export interface Document {
  id: string;
  fileName: string;
  documentType: DocumentDocumentType;
  status: DocumentStatus;
  uploadedAt: string;
  confidence: number;
  extracted: ExtractedFields;
}