# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Tax Sathi API for India income-tax filing preparation
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}