# lib/api-zod/src/generated/types/regimeResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Tax Sathi API for India income-tax filing preparation
 * OpenAPI spec version: 0.1.0
 */

export interface RegimeResult {
  taxableIncome: number;
  baseTax: number;
  cess: number;
  totalTax: number;
  effectiveRate: number;
  rebateApplied?: boolean;
  deductions?: number;
}