# lib/api-zod/src/generated/types/taxCalculationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Tax Sathi API for India income-tax filing preparation
 * OpenAPI spec version: 0.1.0
 */

export interface TaxCalculationInput {
  /** @minimum 0 */
  grossSalary: number;
  /** @minimum 0 */
  otherIncome?: number;
  /** @minimum 0 */
  deductions80c?: number;
  /** @minimum 0 */
  deductions80d?: number;
  /** @minimum 0 */
  deductions80g?: number;
  /** @minimum 0 */
  homeLoanInterest?: number;
  /** @minimum 0 */
  hraReceived?: number;
  /** @minimum 0 */
  rentPaid?: number;
  /** @minimum 0 */
  basicSalary?: number;
  isMetro?: boolean;
  /** @minimum 0 */
  age?: number;
}