# lib/api-zod/src/generated/types/taxSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Tax Sathi API for India income-tax filing preparation
 * OpenAPI spec version: 0.1.0
 */
import type { RegimeResult } from './regimeResult';
import type { TaxSummaryRecommendedRegime } from './taxSummaryRecommendedRegime';

export interface TaxSummary {
  grossIncome: number;
  newRegime: RegimeResult;
  oldRegime: RegimeResult;
  recommendedRegime: TaxSummaryRecommendedRegime;
  savings: number;
  calculatedAt: string;
}